+++ app/views/homes/index.html.erb
@@ -6,5 +6,5 @@
 </div>
 
 
-<%= javascript_pack_tag 'starbucks' %>
-<%= stylesheet_pack_tag 'starbucks' %>
+<%= javascript_pack_tag 'index' %>
+<%= stylesheet_pack_tag 'index' %>
