+++ config/routes.rb
@@ -1,6 +1,8 @@
 Rails.application.routes.draw do
   root to: 'homes#index'
 
+  get '/starbucks',   to: 'homes#index'
+
   namespace :api, {format: 'json'} do
     namespace :v1 do
       resources :starbucks, only: [:index]
